refactor: subshipment util
